requirements.txt
rfdetr
gradio>=4.0.0
huggingface_hub
opencv-python-headless
numpy
torch
torchvision
accelerate==1.13.0
aiofiles==24.1.0      
albucore==0.0.23      
albumentations==1.4.24
annotated-doc==0.0.4  
annotated-types==0.7.0
anyio==4.12.1
brotli==1.2.0
certifi==2026.2.25    
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
exceptiongroup==1.3.1
fastapi==0.135.1
ffmpy==1.0.0
filelock==3.20.0
filetype==1.2.0
fonttools==4.62.0
fsspec==2025.12.0
gradio==6.9.0
gradio_client==2.3.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.7
Jinja2==3.1.6
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84
orjson==3.11.7
packaging==26.0
pandas==2.3.3
peft==0.18.1
pi_heif==1.3.0
pillow==12.0.0
pillow-avif-plugin==1.5.5
psutil==7.2.2
pycocotools==2.0.11
pydantic==2.12.5
pydantic_core==2.41.5
pyDeprecate==0.5.0
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rf100vl==1.1.0
rfdetr==1.5.2
rich==14.3.3
roboflow==1.2.16
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
starlette==0.52.1
stringzilla==4.6.0
supervision==0.27.0.post1
sympy==1.14.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
transformers==4.57.6
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
yt-dlp==2026.3.3